feat: updated admin visualization workflow to unaviable season when they dont exists !!
